remove providers directory
