+++ src/menus/mod.rs
@@ -1,11 +1,13 @@
 pub mod id;
-pub mod providers;
+pub mod provider;
+pub mod skolmaten;
 
 use chrono::NaiveDate;
-use providers::skolmaten::days::SkolmatenMeal;
 use serde::Serialize;
+use skolmaten::days::SkolmatenMeal;
 
-use self::id::MenuID;
+use self::{id::MenuID, provider::Provider};
+use crate::errors::Result;
 
 #[derive(Serialize, Debug)]
 pub struct Meal {
@@ -37,5 +39,23 @@ impl Menu {
             id,
             title: title.to_owned(),
         }
+    }
+}
+
+pub async fn list_menus() -> Result<Vec<Menu>> {
+    let menus = skolmaten::list_menus().await?;
+
+    Ok(menus)
+}
+
+pub async fn query_menu(menu_id: &MenuID) -> Result<Menu> {
+    match menu_id.provider {
+        Provider::Skolmaten => skolmaten::query_menu(&menu_id.local_id).await,
+    }
+}
+
+pub async fn list_days(menu_id: &MenuID) -> Result<Vec<Day>> {
+    match menu_id.provider {
+        Provider::Skolmaten => skolmaten::list_days(&menu_id.local_id).await,
     }
 }
